% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringb-package.R
\docType{package}
\name{about_stringb}
\alias{about_stringb}
\alias{stringb}
\alias{stringb-package}
\title{Drop-in replacements for base R string functions powered by stringi}
\description{
\pkg{stringi} is a mature R package for fast, correct, consistent,
and convenient string/text manipulation.
It gives predictable results on every platform, in each locale,
and under any native character encoding.

\bold{Keywords}: R, text processing, character strings,
internationalisation, localisation, ICU, ICU4C, i18n, l10n, Unicode.

\bold{Homepage}: \url{https://stringb.gagolewski.com/}

\bold{License}: GNU General Public License version 2 or later.
}
\references{
\emph{\pkg{stringi} Package homepage}, \url{https://stringi.gagolewski.com/}

\emph{ICU -- International Components for Unicode}, \url{http://site.icu-project.org/}

\emph{The Unicode Consortium}, \url{https://home.unicode.org/}
}
\author{
Marek Gagolewski
}
