% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartr.R
\name{strtrans}
\alias{strtrans}
\alias{chartr}
\alias{tolower}
\alias{toupper}
\alias{casefold}
\title{Transliteration and Other Text Transforms}
\usage{
strtrans(x, transform)

chartr(old, new, x)

tolower(x)

toupper(x)

casefold(x, upper = NA)
}
\arguments{
\item{x}{character vector (or an object coercible to)}

\item{transform}{a single string with ICU general transform
specifier, see \code{\link[stringi]{stri_trans_list}}.}

\item{old}{a single string}

\item{new}{a single string, preferably of the same length as \code{old}}

\item{upper}{single logical value; switches between case folding
(the default, \code{NA}), lower-, and upper-case.}
}
\value{
A character vector (in UTF-8).

These functions preserve most attributes of \code{x}.
Their base R counterparts drop all the attributes if not fed with character
vectors.
}
\description{
Translate characters, including case mapping and folding,
script to script conversion, and Unicode normalisation.
}
\details{
Unlike their base R counterparts, the new \code{tolower} and
\code{toupper} are locale-sensitive;
see \code{stri_trans_tolower}.

The base \code{casefold} simply dispatches to
\code{tolower} or \code{toupper}
'for compatibility with S-PLUS' (that was only crucial long time ago).
The version implemented here, by default, performs the true case folding,
whose purpose is to make two pieces of text that differ only in case
identical, see \code{stri_trans_casefold}.

The new \code{chartr} is a wrapper for
\code{\link[stringi]{stri_trans_char}}.
Contrary to the base \code{\link[base]{chartr}}, it always generates
a warning when \code{old} and \code{new} are of different lengths.


A new function \code{strtrans} applies ICU general transforms,
see \code{\link[stringi]{stri_trans_general}}.
}
\examples{
strtrans(strcat(letters_bf), "Any-NFKD; Any-Upper")
strtrans(strcat(letters_bb[1:6]), "Any-Hex/C")
strtrans(strcat(letters_greek), "Greek-Latin")

toupper(letters_greek)
tolower(LETTERS_GREEK)

base::toupper("gro\u00DF")
stringx::toupper("gro\u00DF")

casefold("gro\u00DF")

x <- as.matrix(c(a="\u00DFpam ba\U0001D554on spam", b=NA))
base::chartr("\u00DF\U0001D554aba", "SCXBA", x)
stringx::chartr("\u00DF\U0001D554aba", "SCXBA", x)

}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}
}
