% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{digits_dec}
\alias{digits_dec}
\alias{digits_hex}
\alias{letters_greek}
\alias{LETTERS_GREEK}
\alias{letters_bb}
\alias{LETTERS_BB}
\alias{letters_cal}
\alias{LETTERS_CAL}
\alias{letters_frak}
\alias{LETTERS_FRAK}
\alias{letters_bf}
\alias{LETTERS_BF}
\title{Character Constants}
\format{
Decimal digits

Hexadecimal digits

Greek letters (lower case)

Greek letters (upper case)

Blackboard bold English letters (lower case)

Blackboard bold English letters (upper case)

Calligraphy (script) English letters (lower case)

Calligraphy (script) English letters (upper case)

Fraktur English letters (lower case)

Fraktur English letters (upper case)

Bold English letters (lower case)

Bold English letters (upper case)
}
\usage{
digits_dec

digits_hex

letters_greek

LETTERS_GREEK

letters_bb

LETTERS_BB

letters_cal

LETTERS_CAL

letters_frak

LETTERS_FRAK

letters_bf

LETTERS_BF
}
\description{
Letters and digits sets complementing the built-in
\code{LETTERS} and \code{letters}, see \link[base]{Constants}.

Note: calling, e.g., \code{\link{tolower}} on \code{LETTERS_FRAK}
in the current version of \pkg{ICU} does not currently yield
\code{letters_frak}.
}
\examples{
letters_bb
letters_bf
letters_cal
letters_frak
letters_greek
LETTERS_BB
LETTERS_BF
LETTERS_CAL
LETTERS_FRAK
LETTERS_GREEK
digits_dec
digits_hex

}
\keyword{datasets}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}
}
