% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregexpr.R
\name{regexpr2}
\alias{regexpr2}
\alias{gregexpr2}
\alias{regexec2}
\alias{gregexec2}
\alias{regexpr}
\alias{gregexpr}
\alias{regexec}
\alias{gregexec}
\title{Detect Pattern Occurrences}
\usage{
regexpr2(x, pattern, ..., ignore.case = FALSE, fixed = FALSE)

gregexpr2(x, pattern, ..., ignore.case = FALSE, fixed = FALSE)

regexec2(x, pattern, ..., ignore.case = FALSE, fixed = FALSE)

gregexec2(x, pattern, ..., ignore.case = FALSE, fixed = FALSE)

regexpr(
  pattern,
  x = text,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  text
)

gregexpr(
  pattern,
  x = text,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  text
)

regexec(
  pattern,
  x = text,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  text
)

gregexec(
  pattern,
  x = text,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  text
)
}
\arguments{
\item{x}{character vector whose elements are to be examined}

\item{pattern}{character vector of nonempty search patterns}

\item{...}{further arguments to \code{\link[stringi]{stri_split}},
e.g., \code{omit_empty}, \code{locale}, \code{dotall}}

\item{ignore.case}{single logical value; indicates whether matching
should be case-insensitive}

\item{fixed}{single logical value;
\code{FALSE} for matching with regular expressions
    (see \link[stringi]{about_search_regex});
\code{TRUE} for fixed pattern matching
    (\link[stringi]{about_search_fixed});
\code{NA} for the Unicode collation algorithm
    (\link[stringi]{about_search_coll})}

\item{perl, useBytes}{not used (with a warning if
attempting to do so) [DEPRECATED]}

\item{text}{alias to the \code{x} argument [DEPRECATED]}
}
\value{
Returns a list of character vectors representing the identified tokens.
}
\description{
... g stands for 'global'=all
}
\details{
This function is fully vectorised with respect to both arguments.

For splitting text into 'characters' (grapheme clusters), words,
or sentences, use \code{\link[stringi]{stri_split_boundaries}} instead.
}
\section{Differences from Base R}{

Replacements for base \code{\link[base]{strsplit}}
implemented with \code{\link[stringi]{stri_split}}.

\itemize{
\item there are inconsistencies between the argument order and naming
    in \code{\link[base]{grepl}}, \code{\link[base]{strsplit}},
    and \code{\link[base]{startsWith}} (amongst others); e.g.,
    where the needle can precede the haystack, the use of the forward
    pipe operator, \code{\link[base]{|>}}, is less convenient
    \bold{[fixed here]}
\item base R implementation is not portable as it is based on
    the system PCRE or TRE library
    (e.g., some Unicode classes may not be available or matching thereof
    can depend on the current \code{LC_CTYPE} category
    \bold{[fixed here]}
\item not suitable for natural language processing
    \bold{[fixed here -- use \code{fixed=NA}]}
\item two different regular expression libraries are used
    (and historically, ERE was used in place of TRE)
    \bold{[here, \pkg{ICU} Java-like regular expression engine
    is only available, hence the \code{perl} argument has no meaning]}
\item not vectorised w.r.t. \code{pattern}
    \bold{[fixed here]}
\item ...
}
}

\examples{
# ...

}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}

Related function(s): \code{\link{paste}}, \code{\link{nchar}},
    \code{\link{strsplit}}, \code{\link{gsub}}, \code{\link{substr}},
    \code{\link{grepl}}
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
