% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep.R
\name{grep}
\alias{grep}
\alias{grepl}
\alias{regexpr}
\alias{gregexpr}
\alias{regexec}
\alias{gregexec}
\title{Detect Pattern Occurrences}
\usage{
grep(
  pattern,
  x,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  value = FALSE,
  invert = FALSE,
  perl = FALSE,
  useBytes = FALSE
)

grepl(
  pattern,
  x,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE
)

regexpr(
  pattern,
  x = text,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  text
)

gregexpr(
  pattern,
  x = text,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  text
)

regexec(
  pattern,
  x = text,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  text
)

gregexec(
  pattern,
  x = text,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  text
)
}
\arguments{
\item{pattern}{character vector of nonempty search patterns}

\item{x}{character vector whose elements are to be examined}

\item{...}{further arguments to \code{\link[stringi]{stri_split}},
e.g., \code{omit_empty}, \code{locale}, \code{dotall}}

\item{ignore.case}{single logical value; indicates whether matching
should be case-insensitive}

\item{fixed}{single logical value;
\code{FALSE} for matching with regular expressions
    (see \link[stringi]{about_search_regex});
\code{TRUE} for fixed pattern matching
    (\link[stringi]{about_search_fixed});
\code{NA} for the Unicode collation algorithm
    (\link[stringi]{about_search_coll})}

\item{perl, useBytes}{not used (with a warning if
attempting to do so) [DEPRECATED]}
}
\value{
Returns a list of character vectors representing the identified tokens.
}
\description{
...
}
\details{
This function is fully vectorised with respect to both arguments.

For splitting text into 'characters' (grapheme clusters), words,
or sentences, use \code{\link[stringi]{stri_split_boundaries}} instead.
}
\section{Differences from Base R}{

Replacements for base \code{\link[base]{strsplit}}
implemented with \code{\link[stringi]{stri_split}}.

\itemize{
\item base R implementation is not portable as it is based on
    the system PCRE or TRE library
    (e.g., some Unicode classes may not be available or matching thereof
    can depend on the current \code{LC_CTYPE} category
    \bold{[fixed here]}
\item not suitable for natural language processing
    \bold{[fixed here -- use \code{fixed=NA}]}
\item two different regular expression libraries are used
    (and historically, ERE was used in place of TRE)
    \bold{[here, \pkg{ICU} Java-like regular expression engine
    is only available, hence the \code{perl} argument has no meaning]}
\item there are inconsistencies between the argument order and naming
    in \code{\link[base]{grepl}}, \code{\link[base]{strsplit}},
    and \code{\link[base]{startsWith}} (amongst others); e.g.,
    where the needle can precede the haystack, the use of the forward
    pipe operator \code{|>} is less convenient
    \bold{[............fixed here]}
\item \code{\link[base]{grepl}} also features the \code{ignore.case} argument
    \bold{[added here]}
\item if \code{split} is a zero-length vector, it is treated as \code{""},
    which extracts individual code points (which is not the best idea
    for natural language processing tasks)
    \bold{[empty search patterns are not supported here, zero-length vectors
    are propagated correctly]}
\item last empty token is removed from the output, but first is not
    \bold{[fixed here -- see also the \code{omit_empty} argument]}
\item missing values in \code{split} are not propagated correctly
    \bold{[fixed here]}
\item partial recycling without the usual warning, not fully vectorised
    w.r.t. the \code{split} argument
    \bold{[fixed here]}
\item only the \code{names} attribute of \code{x} is preserved
    \bold{[fixed here]}
}
}

\examples{
# ...

}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}

Related function(s): \code{\link{paste}}, \code{\link{nchar}},
    \code{\link{strsplit}}, \code{\link{gsub}}, \code{\link{substr}}
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
