% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsub.R
\name{sub}
\alias{sub}
\alias{gsub}
\title{Replace Pattern Occurrences}
\usage{
sub(
  pattern,
  replacement,
  x,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE
)

gsub(
  pattern,
  replacement,
  x,
  ...,
  ignore.case = FALSE,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{pattern}{character vector of nonempty search patterns}

\item{replacement}{character vector with the corresponding replacement
strings}

\item{x}{character vector with strings whose chunks are to be modified}

\item{...}{further arguments to \code{\link[stringi]{stri_replace}},
e.g., \code{omit_empty}, \code{locale}, \code{dotall}}

\item{ignore.case}{single logical value; indicates whether matching
should be case-insensitive}

\item{fixed}{single logical value;
\code{FALSE} for matching with regular expressions
    (see \link[stringi]{about_search_regex});
\code{TRUE} for fixed pattern matching
    (\link[stringi]{about_search_fixed});
\code{NA} for the Unicode collation algorithm
    (\link[stringi]{about_search_coll})}

\item{perl, useBytes}{not used (with a warning if
attempting to do so) [DEPRECATED]}

\item{text}{alias to the \code{x} argument [DEPRECATED]}
}
\value{
Both functions return a character vector.
Attributes are copied from the longest inputs.
}
\description{
Splits each string into chunks delimited by occurrences of a given pattern.
}
\details{
These functions are fully vectorised with respect to \code{pattern},
\code{replacement}, and \code{x}.

For splitting text into 'characters' (grapheme clusters), words,
or sentences, use \code{\link[stringi]{stri_split_boundaries}} instead.
}
\section{Differences from Base R}{

Replacements for base \code{\link[base]{strsplit}}
implemented with \code{\link[stringi]{stri_split}}.

\itemize{
\item base R implementation is not portable as it is based on
    the system PCRE or TRE library
    (e.g., some Unicode classes may not be available or matching thereof
    can depend on the current \code{LC_CTYPE} category
    \bold{[fixed here]}
\item not suitable for natural language processing
    \bold{[fixed here -- use \code{fixed=NA}]}
\item two different regular expression libraries are used
    (and historically, ERE was used in place of TRE)
    \bold{[here, \pkg{ICU} Java-like regular expression engine
    is only available, hence the \code{perl} argument has no meaning]}
\item there are inconsistencies between the argument order and naming
    in \code{\link[base]{grepl}}, \code{\link[base]{strsplit}},
    and \code{\link[base]{startsWith}} (amongst others); e.g.,
    where the needle can precede the haystack, the use of the forward
    pipe operator \code{|>} is less convenient
    \bold{[...........fixed here]}
\item \code{\link[base]{grepl}} also features the \code{ignore.case} argument
    \bold{[added here]}
\item if \code{split} is a zero-length vector, it is treated as \code{""},
    which extracts individual code points (which is not the best idea
    for natural language processing tasks)
    \bold{[empty search patterns are not supported here, zero-length vectors
    are propagated correctly]}
\item last empty token is removed from the output, but first is not
    \bold{[fixed here -- see also the \code{omit_empty} argument]}
\item missing values in \code{split} are not propagated correctly
    \bold{[fixed here]}
\item partial recycling without the usual warning, not fully vectorised
    w.r.t. the \code{split} argument
    \bold{[fixed here]}
\item only the \code{names} attribute of \code{x} is preserved
    \bold{[fixed here]}
\item ...different syntax....
}
}

\examples{
# ...

}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}

Related function(s): \code{\link{paste}}, \code{\link{nchar}},
    \code{\link{grep}}, \code{\link{strsplit}}, \code{\link{substr}}
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
