% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchar.R
\name{nchar}
\alias{nchar}
\alias{nzchar}
\title{Get Length or Width of Strings}
\usage{
nchar(x, type = "chars", allowNA = FALSE, keepNA = TRUE)

nzchar(x, keepNA = TRUE)
}
\arguments{
\item{x}{character vector or an object coercible to}

\item{type}{\code{"chars"} gives the number of code points,
\code{"width"} estimates the string width,
\code{"bytes"} computes the number of bytes}

\item{allowNA}{not used}

\item{keepNA}{if \code{FALSE}, missing values will
be treated as \code{"NA"} strings; otherwise, the corresponding outputs
will be missing as well}
}
\value{
\code{nchar} returns an integer vector.

\code{nzchar} returns a logical vector, where \code{TRUE} indicates
that the corresponding string is of non-zero length (i.e., non-empty).
}
\description{
\code{nchar} computes the number of code points, bytes used, or
estimated total width of strings in a character vector.
\code{nzchar} indicates which strings are empty.
}
\details{
Replacement for base \code{\link[base]{nchar}} and \code{\link[base]{nzchar}}
implemented with \code{\link[stringi]{stri_length}},
\code{\link[stringi]{stri_width}},
\code{\link[stringi]{stri_numbytes}},
and \code{\link[stringi]{stri_isempty}}

Inconsistencies in base R
and the way we have addressed them here:

\itemize{
\item \code{keepNA} does not default to \code{TRUE}, and hence
    missing values are treated as \code{"NA"} strings
    \bold{[fixed here]};
\item TODO: width... emoji of width 2, combining characters and modifiers (e.g., skin tones)
   recognised properly \bold{[fixed here]};
\item TODO: attributes?
}
}
\examples{
# dark skin tone flexed biceps:
base::nchar("\U0001F4AA\U0001F3FF", "width")
stringx::nchar("\U0001F4AA\U0001F3FF", "width")


}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}
}
