% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{xtfrm}
\alias{xtfrm}
\alias{sort.character}
\alias{xtfrm.character}
\title{Compare Strings...............}
\usage{
xtfrm(x, ...)

\method{sort}{character}(
  x,
  decreasing = FALSE,
  na.last = NA,
  locale = NULL,
  strength = 3L,
  alternate_shifted = FALSE,
  french = FALSE,
  uppercase_first = NA,
  case_level = FALSE,
  normalisation = FALSE,
  numeric = FALSE
)

\method{xtfrm}{character}(
  x,
  locale = NULL,
  strength = 3L,
  alternate_shifted = FALSE,
  french = FALSE,
  uppercase_first = NA,
  case_level = FALSE,
  normalisation = FALSE,
  numeric = FALSE
)
}
\arguments{
\item{x}{character vector whose elements are to be sorted}

\item{decreasing}{single logical value; if \code{FALSE}, the ordering
is nondecreasing (weakly increasing)}

\item{na.last}{single logical value; if \code{TRUE}, then missing values
are placed at the end; if \code{FALSE}, they are put at the beginning;
if \code{NA}, then they are removed from the output whatsoever.}

\item{locale}{see \code{\link[stringi]{stri_opts_collator}}}

\item{strength}{see \code{\link[stringi]{stri_opts_collator}}}

\item{alternate_shifted}{see \code{\link[stringi]{stri_opts_collator}}}

\item{french}{see \code{\link[stringi]{stri_opts_collator}}}

\item{uppercase_first}{see \code{\link[stringi]{stri_opts_collator}}}

\item{case_level}{see \code{\link[stringi]{stri_opts_collator}}}

\item{normalisation}{see \code{\link[stringi]{stri_opts_collator}}}

\item{numeric}{see \code{\link[stringi]{stri_opts_collator}}}
}
\value{
...
}
\description{
...
}
\details{
.....Replacements for base \link[base]{Comparison} operators
implemented with \code{\link[stringi]{stri_rank}}.
\code{\link[stringi]{stri_sort}}

order, rank - no - show example what to use....

we need to overload the built-in C-level dispatcher
because it only supports 1 argument and does not recognise
the method overloaded for objects of class 'character'.....

Inconsistencies in base R and the way we have addressed them here:

\itemize{
\item collation in different locales is difficult and non-portable across
    platforms
    \bold{[fixed here -- using services provided by ICU]}
\item overloading \code{`xtfrm.character`} has no effect in R, because S3
    method dispatch is done internally with hard-coded support for
    character arguments. Thus, we needed to replace the generic
    \code{`xtfrm`} with the one that calls \code{\link[base]{UseMethod}}
    \bold{[fixed here]}
\item order has na.last=TRUE, sort removes NA.....
\item \code{`xtfrm`} does not support customisation of the linear ordering
    relation it is based upon
    \bold{[fixed by introducing \code{...} argument to the generic]}
}
}
\examples{
# ...
# order, rank examples too also with dfs...
library("stringx")
x <- c("1", "10", "11", "99", "100", "101", "1000")
base::sort.default(x)
sort(x, numeric=TRUE)  # calls stringx::sort.character
xtfrm(x, numeric=TRUE)


}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}

Related function(s): \code{\link{strcoll}}
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
