% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startswith.R
\name{startsWith}
\alias{startsWith}
\alias{endsWith}
\title{Detect Pattern Occurrences at Start or End of Strings}
\usage{
startsWith(x, prefix, ignore.case = FALSE)

endsWith(x, suffix, ignore.case = FALSE)
}
\arguments{
\item{x}{character vector whose elements are to be examined}

\item{prefix, suffix}{character vectors with fixed (literal) patterns
to search for}

\item{ignore.case}{single logical value}
}
\value{
Each function returns a logical vector, indicating whether a pattern
match has been detected or not.
}
\description{
Determines if a string starts or ends if with a match to a specified
fixed pattern.
}
\details{
Replacements for base \code{\link[base]{startsWith}}
and \code{\link[base]{endsWith}}
implemented with \code{\link[stringi]{stri_startswith_fixed}}
and \code{\link[stringi]{stri_endswith_fixed}}

For matching with regular expressions, see \code{\link{grepl}}
with patterns like \code{"^prefix"} and \code{"suffix$"}.

Vectorised with respect to both arguments.

Inconsistencies in/differences from base R:

\itemize{
\item note that other pattern matching functions have a different argument
    order, where the needle precedes the haystack
    \bold{[not fixed here]}
\item \code{\link[base]{grepl}} also features the \code{ignore.case} argument
    \bold{[added here]}
\item partial recycling without the usual warning
    \bold{[fixed here]}
\item no attributes preserved whatsoever
    \bold{[fixed here]}
}
}
\examples{
stringx::startsWith("ababa", c("a", "ab", "aba", "baba", NA))
outer(
    c("aba", "abb", "abc", "baba", "bac"),
    c("A", "B", "C"),
    stringx::endsWith,
    ignore.case=TRUE
)

}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}

Related function(s): \code{\link{grepl}}, \code{\link{substring}}
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
