% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strptime.R
\name{strptime}
\alias{strptime}
\alias{strftime}
\title{Date-time Parsing and Formatting}
\usage{
strptime(
  x,
  format = "\%Y-\%m-\%dT\%H:\%M:\%S\%z",
  tz = "",
  lenient = FALSE,
  locale = NULL
)

strftime(
  x,
  format = "\%Y-\%m-\%dT\%H:\%M:\%S\%z",
  tz = "",
  usetz = FALSE,
  ...,
  locale = NULL
)
}
\arguments{
\item{x}{object to be converted}

\item{format}{character vector of date-time format specifiers,
see \code{\link[stringi]{stri_datetime_fstr}};
e.g., \code{"\%Y-\%m-\%d"} or \code{"datetime_full"};
the default conforms to the ISO 8601 guideline,
e.g., '2015-12-31T23:59:59+0100'}

\item{tz}{\code{NULL} or \code{''} for the default time zone
(see \code{\link[stringi]{stri_timezone_get}})
or a single string with a timezone identifier,
see \code{\link[stringi]{stri_timezone_list}}}

\item{lenient}{single logical value; should date/time parsing be lenient?}

\item{locale}{\code{NULL} or \code{''} for the default locale
(see \code{\link[stringi]{stri_locale_get}})
or a single string with a locale identifier,
see \code{\link[stringi]{stri_locale_list}}}

\item{usetz}{....}

\item{...}{not used}
}
\value{
\code{strftime} returns a character vector (in UTF-8).

\code{strptime} returns an object of class \code{\link{POSIXlt}}.
}
\description{
\code{strptime} parses strings representing date-time data
and converts it to a date-time object.
\code{strftime} formats a date-time object and outputs it as a
character vector.
}
\details{
Replacements for base \code{\link[base]{strptime}}
and \code{\link[base]{strftime}} implemented with
\code{\link[stringi]{stri_datetime_parse}} and
\code{\link[stringi]{stri_datetime_format}}.

Inconsistencies/limitations in base R and the way we have addressed them:

\itemize{
\item formatting/parsing date-time in different locales and calendars
    is difficult and non-portable across platforms
    \bold{[fixed here - using services provided by ICU]};
\item default format not conforming to ISO 8601, in particular not
    displaying the current time zone
    \bold{[fixed here]};
\item partial recycling with no warning
    \bold{[fixed here]};
... format 2nd arg (but sprintf has as 1st); undocumented default arg
}
}
\examples{
f <- c("date_full", "\%Y-\%m-\%d", "date_relative_short", "datetime_full")
stringx::strftime(Sys.time(), f)  # current default locale
stringx::strftime(Sys.time(), f, locale="de_DE")
stringx::strftime(Sys.time(), "date_short", locale="en_IL@calendar=hebrew")


}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}

Related function(s): \code{\link{sprintf}}
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
