% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr.R
\name{substr}
\alias{substr}
\alias{substrl}
\alias{substring}
\alias{substr<-}
\alias{substrl<-}
\alias{substring<-}
\title{Extract or Replace Substrings}
\usage{
substr(x, start = 1L, stop = -1L)

substrl(x, start = 1L, length)

substring(text, first = 1L, last = -1L)

substr(x, start = 1L, stop = -1L) <- value

substrl(x, start = 1L, length) <- value

substring(text, first = 1L, last = -1L) <- value
}
\arguments{
\item{x, text}{character vector
whose parts are to be extracted/replaced}

\item{start, first}{numeric vector giving the start indexes;
e.g., 1 points to the first code point, -1 to the last}

\item{stop, last}{numeric vector giving the end indexes (inclusive);
as with \code{start}, for negative indexes, counting starts at the end
of each string; note that if the start position is farther than the
end position, this indicates an empty substring therein (see Examples)}

\item{length}{numeric vector giving the substring lengths}

\item{value}{character vector defining the replacements strings}
}
\value{
\code{substr} and \code{substrl} return a character vector (in UTF-8).
Their replacement versions modify \code{x} 'in-place' (see Examples).

The attributes are copied from the longest arguments (similarly
as in the case of binary operators).
}
\description{
\code{substr} and \code{substrl} extract
contiguous parts of given character strings.
The former operates based on start and end positions
while the later is fed with start position and substring length.

Their replacement versions allows for substituting them with new content.
}
\details{
Not to be confused with \code{\link{sub}}.

\code{substring} is a [DEPRECATED] synonym for \code{substr}.

Note that these functions can break some meaningful Unicode code point
sequences, e.g., when inputs are not normalised. For extracting
initial parts of strings based on character width, see \code{\link{strtrim}}.
}
\section{Differences from Base R}{

Replacement for base \code{\link[base]{substr}}
and \code{\link[base]{substring}}
implemented with \code{\link[stringi]{stri_sub}}.

\itemize{
\item \code{substring} is "for compatibility with S", but this should
    no longer matter
    \bold{[here, \code{substring} is equivalent to \code{substr}; in a
    future version, using the former may result in a warning]}
\item \code{substr} is not vectorised with respect to all the arguments
    (and \code{substring} is not fully vectorised wrt \code{value})
    \bold{[fixed here]}
\item not all attributes are taken form the longest of the inputs
    \bold{[fixed here]}
\item partial recycling with no warning
    \bold{[fixed here]}
\item if the replacement string of different length than the chunk
    being substituted, then
    \bold{[fixed here]}
\item negative indexes are silently treated as 1
    \bold{[changed here -- negative indexes count from the end of the string]}
\item replacement of different length than the extracted substring
    never changes the length of the string
    \bold{[changed here -- output length is input length minus
    length of extracted plus length of replacement]}
\item \code{\link{gregexpr}} (amongst others) return start positions
    and lengths of matches, but base \code{substr} only uses
    start and end
    \bold{[fixed by introducing \code{substrl}]}
}
}

\examples{
x <- "spam, spam, bacon, and spam"
base::substr(x, c(1, 13), c(4, 17))
base::substring(x, c(1, 13), c(4, 17))
substr(x, c(1, 13), c(4, 17))
substrl(x, c(1, 13), c(4, 5))

# replacement function used as an ordinary one - return a copy of x:
base::`substr<-`(x, 1, 4, value="jam")
`substr<-`(x, 1, 4, value="jam")
base::`substr<-`(x, 1, 4, value="porridge")
`substr<-`(x, 1, 4, value="porridge")

# replacement function modifying x in-place:
substr(x, 1, 4) <- "eggs"
substr(x, 1, 0) <- "porridge, "        # prepend (start<stop)
substr(x, nchar(x)+1) <- " every day"  # append (start<stop)
print(x)


}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}

Related function(s): \code{\link{strtrim}}, \code{\link{nchar}},
   \code{\link{startsWith}}, \code{\link{endsWith}},
   \code{\link{gregexpr}}

See also \code{\link[stringi]{stri_sub_all}} for replacing
multiple substrings within individual strings.
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
