% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_stringx.R
\docType{package}
\name{about_stringx}
\alias{about_stringx}
\alias{stringx}
\alias{stringx-package}
\title{Drop-in Replacements for Base String Functions Powered by Stringi}
\description{
\pkg{stringx} reimplements the built-in R string processing functions
based on \pkg{stringi} -- a mature R package for fast, correct, consistent,
and convenient text manipulation. Thanks to the \pkg{ICU} library, we obtain
predictable results on every platform, in each locale, and under any
native character encoding.

\bold{Keywords}: R, text processing, character strings,
internationalisation, localisation, ICU, ICU4C, i18n, l10n, Unicode

\bold{License}: GNU General Public License version 2 or later
}
\references{
\emph{\pkg{stringi} Package homepage},
\url{https://stringi.gagolewski.com/}

\emph{ICU -- International Components for Unicode},
\url{https://icu.unicode.org/}

\emph{The Unicode Consortium},
\url{https://home.unicode.org/}
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}
}
