% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startswith.R
\name{startsWith}
\alias{startsWith}
\alias{endsWith}
\title{Detect Pattern Occurrences at Start or End of Strings}
\usage{
startsWith(x, prefix, fixed = TRUE, ignore.case = FALSE, ...)

endsWith(x, suffix, fixed = TRUE, ignore.case = FALSE, ...)
}
\arguments{
\item{x}{character vector whose elements are to be examined}

\item{prefix, suffix}{character vectors with patterns to search for}

\item{fixed}{single logical value;
\code{TRUE} for fixed pattern matching
    (see \link[stringi]{about_search_fixed});
\code{NA} for the Unicode collation algorithm
    (\link[stringi]{about_search_coll});
\code{FALSE} is not supported -- use \code{\link{grepl}} instead}

\item{ignore.case}{single logical value; indicates whether matching
should be case-insensitive}

\item{...}{further arguments to \code{\link[stringi]{stri_startswith}}
and \code{\link[stringi]{stri_endswith}}}
}
\value{
Each function returns a logical vector, indicating whether a pattern
match has been detected or not.
}
\description{
Determines if a string starts or ends with a match to a specified
fixed pattern.
}
\details{
Both functions are fully vectorised with respect to both arguments.

For matching with regular expressions, see \code{\link{grepl}}
with patterns like \code{"^prefix"} and \code{"suffix$"}.
}
\section{Differences from base R}{

Replacements for base \code{\link[base]{startsWith}}
and \code{\link[base]{endsWith}}
implemented with \code{\link[stringi]{stri_startswith}}
and \code{\link[stringi]{stri_endswith}}.

\itemize{
\item \code{\link[base]{grepl}} and some other pattern matching functions
    have a different argument order, where the needle precedes the haystack
    \bold{[not fixed here]}
\item \code{\link[base]{grepl}} also features the \code{ignore.case} argument
    \bold{[added here]}
\item partial recycling without the usual warning
    \bold{[fixed here]}
\item no attributes preserved whatsoever
    \bold{[fixed here]}
\item not suitable for natural language processing
    \bold{[fixed here -- use \code{fixed=NA}]}
}
}

\examples{
stringx::startsWith("ababa", c("a", "ab", "aba", "baba", NA))
outer(
    c("aba", "abb", "abc", "baba", "bac"),
    c("A", "B", "C"),
    stringx::endsWith,
    ignore.case=TRUE
)
x <- c("Mario", "mario", "M\u00E1rio", "M\u00C1RIO", "Mar\u00EDa", "Rosario")
x[stringx::startsWith(x, "mario", ignore.case=TRUE)]
x[stringx::startsWith(x, "mario", fixed=NA, strength=1L)]

}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}

Related function(s): \code{\link{grepl}}, \code{\link{substr}}
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
