% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr.R
\name{substr}
\alias{substr}
\alias{substrl}
\alias{substr<-}
\alias{substrl<-}
\alias{gsubstrl}
\alias{gsubstr<-}
\alias{gsubstrl<-}
\alias{substring}
\alias{substring<-}
\title{Extract or Replace Substrings}
\usage{
substr(x, start = 1L, stop = -1L)

substrl(x, start = 1L, length = attr(start, "match.length"))

substr(x, start = 1L, stop = -1L) <- value

substrl(x, start = 1L, length = attr(start, "match.length")) <- value

gsubstrl(x, start = list(1L), length = lapply(start, attr, "match.length"))

gsubstr(x, start = list(1L), stop = list(-1L)) <- value

gsubstrl(x, start = list(1L), length = lapply(start, attr, "match.length")) <- value

substring(text, first = 1L, last = -1L)

substring(text, first = 1L, last = -1L) <- value
}
\arguments{
\item{x, text}{character vector
whose parts are to be extracted/replaced}

\item{start, first}{numeric vector giving the start indexes;
e.g., 1 denotes the first code point;
negative indexes
count from the end of a string, i.e., -1 is the last character}

\item{stop, last}{numeric vector giving the end indexes (inclusive);
note that if the start position is farther than the
end position, this indicates an empty substring therein (see Examples)}

\item{length}{numeric vector giving the substring lengths;
negative lengths result in an empty string or the corresponding
substring being unchanged}

\item{value}{character vector defining the replacements strings}
}
\value{
\code{substr2} and \code{substrl2} return a character vector (in UTF-8).
Their replacement versions modify \code{x} 'in-place' (see Examples).

The attributes are copied from the longest arguments (similar to
binary operators).
}
\description{
\code{substr} and \code{substrl} extract
contiguous parts of given character strings.
The former operates based on start and end positions
while the latter is fed with substring lengths.

Their replacement versions allow for substituting parts of strings
with new content.
}
\details{
Not to be confused with \code{\link{sub}}.

\code{substring} is a [DEPRECATED] synonym for \code{substr}.

Note that these functions can break some meaningful Unicode code point
sequences, e.g., when inputs are not normalised. For extracting
initial parts of strings based on character width, see \code{\link{strtrim}}.
}
\section{Differences from Base R}{

Replacement for base \code{\link[base]{substr}}
and \code{\link[base]{substring}}
implemented with \code{\link[stringi]{stri_sub}}.

\itemize{
\item \code{substring} is "for compatibility with S", but this should
    no longer matter
    \bold{[here, \code{substring} is equivalent to \code{substr}; in a
    future version, using the former may result in a warning]}
\item \code{substr} is not vectorised with respect to all the arguments
    (and \code{substring} is not fully vectorised wrt \code{value})
    \bold{[fixed here]}
\item not all attributes are taken from the longest of the inputs
    \bold{[fixed here]}
\item partial recycling with no warning
    \bold{[fixed here]}
\item if the replacement string of different length than the chunk
    being substituted, then
    \bold{[fixed here]}
\item negative indexes are silently treated as 1
    \bold{[changed here -- negative indexes count from the end of the string]}
\item replacement of different length than the extracted substring
    never changes the length of the string
    \bold{[changed here -- output length is input length minus
    length of extracted plus length of replacement]}
\item \code{\link{gregexpr}} (amongst others) return start positions
    and lengths of matches, but base \code{substr} only uses
    start and end
    \bold{[fixed by introducing \code{substrl}]}
}
}

\examples{
x <- "spam, spam, bacon, and spam"
base::substr(x, c(1, 13), c(4, 17))
base::substring(x, c(1, 13), c(4, 17))
substr(x, c(1, 13), c(4, 17))
substrl(x, c(1, 13), c(4, 5))

# replacement function used as an ordinary one - return a copy of x:
base::`substr<-`(x, 1, 4, value="jam")
`substr<-`(x, 1, 4, value="jam")
base::`substr<-`(x, 1, 4, value="porridge")
`substr<-`(x, 1, 4, value="porridge")

# replacement function modifying x in-place:
substr(x, 1, 4) <- "eggs"
substr(x, 1, 0) <- "porridge, "        # prepend (start<stop)
substr(x, nchar(x)+1) <- " every day"  # append (start<stop)
print(x)


}
\seealso{
The official online manual of \pkg{stringx} at \url{https://stringx.gagolewski.com/}

Related function(s): \code{\link{strtrim}}, \code{\link{nchar}},
   \code{\link{startsWith}}, \code{\link{endsWith}},
   \code{\link{gregexpr}}

See also \code{\link[stringi]{stri_sub_all}} for replacing
multiple substrings within individual strings.
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
